added confirmation on delete, removed namespace es

+++ src/Commands/ElasticSearch/DeleteIndexCommand.php
@@ -5,6 +5,7 @@ use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Console\Question\ConfirmationQuestion;
 
 use Buonzz\Scalp\ElasticServer;
 
@@ -13,14 +14,22 @@ class DeleteIndexCommand extends Command
     protected function configure()
     {
         $this
-            ->setName('es:delete')
+            ->setName('delete')
             ->setDescription('delete ElasticSearch index');
     }
 
     protected function execute(InputInterface $input, OutputInterface $output)
     {
-        ElasticServer::delete_index();        
-        $output->writeln("Index Deleted!");
+
+        $helper = $this->getHelper('question');
+        $question = new ConfirmationQuestion('Are you sure you want to delete <comment>"'. getenv('DB_NAME') . '</comment>" database on <comment>' .getenv('DB_HOSTNAME') . ':' . getenv('DB_PORT') . '</comment> ? ( default: no ) ', false);
+
+        if ($helper->ask($input, $output, $question)) {
+            ElasticServer::delete_index();        
+            $output->writeln("Index Deleted!");
+        }
+        else
+            $output->writeln("Operation aborted");
     }
 
 }
